+++ MonacoEditorComponent/Monaco/MarkerSeverity.cs
@@ -1,8 +1,5 @@
 ﻿namespace Monaco
 {
-    /// <summary>
-    /// https://microsoft.github.io/monaco-editor/api/enums/monaco.severity.html
-    /// </summary>
     public enum MarkerSeverity
     {
         Hint = 1,