+++ changelog.md
@@ -10,6 +10,8 @@ v0.9 - 01/23/2020
 - Use StandaloneEditorConstructionOptions for Options instead of IEditorConstructionOptions
 - Use ClassName string instead of CssStyle type
 - Snippets need InsertTextRules property set to InsertAsSnippet
+- Use HoverProvider interface instead of original delegate
+- IUri has been renamed to Uri
 
 v0.8.1 - 01/15/2019
 -------------------